+++ shvatka/infrastructure/db/dao/complex/waiver.py
@@ -2,7 +2,8 @@ import typing
 from dataclasses import dataclass
 from typing import Iterable, Sequence
 
-from shvatka.core.interfaces.dal.waiver import WaiverVoteAdder, WaiverVoteGetter, WaiverApprover
+from shvatka.core.interfaces.dal.waiver import WaiverApprover
+from shvatka.core.waiver.adapters import WaiverVoteAdder, WaiverVoteGetter
 from shvatka.core.models import dto
 from shvatka.core.models.enums.played import Played
