+++ shvatka/infrastructure/di/interactors.py
@@ -17,6 +17,7 @@ from shvatka.core.games.adapters import (
 from shvatka.core.interfaces.current_game import CurrentGameProvider
 from shvatka.core.interfaces.dal.game import GameByIdGetter
 from shvatka.core.interfaces.dal.game_play import GamePlayerDao
+from shvatka.core.waiver.adapters import WaiverVoteAdder, WaiverVoteGetter
 from shvatka.core.scenario.interactors import (
     AllGameKeysReaderInteractor,
     GameScenarioTransitionsInteractor,
@@ -25,6 +26,7 @@ from shvatka.core.services.current_game import CurrentGameProviderImpl
 from shvatka.core.services.key import KeyProcessor, TimerProcessor
 from shvatka.core.services.game_play import CheckKeyInteractor
 from shvatka.core.waiver.interactors import WaiversReaderInteractor, AddWaiverVoteInteractor
+from shvatka.infrastructure.db.dao.complex import WaiverVoteAdderImpl, WaiverVoteGetterImpl
 from shvatka.infrastructure.db.dao.complex.game import GameFilesGetterImpl, GamePlayReaderImpl
 from shvatka.infrastructure.db.dao.complex.game import (
     GameFilesGetterImpl,
@@ -89,3 +91,6 @@ class WaiverProvider(Provider):
 
     waivers_reader_interactor = provide(WaiversReaderInteractor)
     add_waiver_vote = provide(AddWaiverVoteInteractor)
+
+    waiver_vote_adder_dao = provide(WaiverVoteAdderImpl, provides=WaiverVoteAdder)
+    waiver_vote_getter_dao = provide(WaiverVoteGetterImpl, provides=WaiverVoteGetter)
